Ignore hostNetwork Pods in PodStore
The PodStore is used to map a Pod IP to a specific Pod resource, taking
into account time as a way to distinguish between Pods when an IP
address is reused by the CNI.

This mapping is not applicable to hostNetwork Pods, as there will
typically be multiple hostNetwork Pods on a given Node, and they will
all "share" the same (host) IP. Therefore, we modify the PodStore
implementation to ignore events for hostNetwork Pods.

Additionally, for the Flow Aggregator, we switch to a filtered informer,
which will ignore hostNetwork Pods at the server level, and we trim Pods
locally to reduce the overall memory footprint.

Signed-off-by: Antonin Bas <antonin.bas@broadcom.com>