Reconcile host-network Pods after agent is restarted (#6944)
This change is to support reconciling the host-network Pods on Windows because
k8s expects to let CNI manage such Pods as long as they are not using
host-process containers.

Antrea received the CmdAdd request for such Pods when they were created, so
they should be included in the Pod reconcile list after agent is restarted.

Signed-off-by: Wenying Dong <wenyingd@vmware.com>